# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminReportResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "target_type",
          "target_id",
          "reason",
          "description",
          "status",
          "created_at"
        ],
        "title": "AdminReportResponse",
        "type": "object"
      },
...
      "BookingResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "deposit_minor": {
            "title": "Deposit Minor",
            "type": "integer"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "fulfillment_method": {
            "title": "Fulfillment Method",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "listing_id": {
            "format": "uuid",
            "title": "Listing Id",
            "type": "string"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "public_number": {
            "title": "Public Number",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "rental_charge_minor": {
            "title": "Rental Charge Minor",
            "type": "integer"
          },
          "renter_user_id": {
            "format": "uuid",
            "title": "Renter User Id",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "public_number",
          "listing_id",
          "renter_user_id",
          "owner_user_id",
          "starts_at",
          "ends_at",
          "quantity",
          "status",
          "fulfillment_method",
          "rental_charge_minor",
          "deposit_minor",
          "currency",
          "created_at"
        ],
        "title": "BookingResponse",
        "type": "object"
      },
...
      "ChallengeConfirmRequest": {
        "properties": {
          "challenge_id": {
            "format": "uuid",
            "title": "Challenge Id",
            "type": "string"
          },
          "code": {
            "pattern": "^\\d{6}$",
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "code"
        ],
        "title": "ChallengeConfirmRequest",
        "type": "object"
      },
      "ConditionReportRequest": {
        "properties": {
          "checklist": {
            "additionalProperties": true,
            "title": "Checklist",
            "type": "object"
          },
          "notes": {
            "default": "",
            "maxLength": 3000,
            "title": "Notes",
            "type": "string"
          },
          "phase": {
            "enum": [
              "handover",
              "return"
            ],
            "title": "Phase",
            "type": "string"
          }
        },
        "required": [
          "phase"
        ],
        "title": "ConditionReportRequest",
        "type": "object"
      },
      "DisputeRequest": {
        "properties": {
          "description": {
            "maxLength": 5000,
            "minLength": 10,
            "title": "Description",
            "type": "string"
          },
          "type": {
            "enum": [
              "item_not_received",
              "item_different",
              "damage",
              "missing_parts",
              "late_return",
              "non_return",
              "payment_disagreement",
              "unsafe_behavior",
              "abusive_communication"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "description"
        ],
        "title": "DisputeRequest",
        "type": "object"
      },
...
      "FulfillmentResponse": {
        "properties": {
          "booking_id": {
            "format": "uuid",
            "title": "Booking Id",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "private_instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Private Instructions"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "booking_id",
          "method",
          "status",
          "scheduled_at",
          "private_instructions"
        ],
        "title": "FulfillmentResponse",
        "type": "object"
      },
      "FulfillmentScheduleRequest": {
        "properties": {
          "private_instructions": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Private Instructions"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          }
        },
        "title": "FulfillmentScheduleRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModerationRequest": {
        "properties": {
          "action": {
            "enum": [
              "approve",
              "remove"
            ],
            "title": "Action",
            "type": "string"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "action",
          "reason"
        ],
        "title": "ModerationRequest",
        "type": "object"
      },
      "PaymentAcknowledgementRequest": {
        "properties": {
          "disagreement": {
            "default": false,
            "title": "Disagreement",
            "type": "boolean"
          }
        },
        "title": "PaymentAcknowledgementRequest",
        "type": "object"
      },
...
      "ReportRequest": {
        "properties": {
          "description": {
            "default": "",
            "maxLength": 3000,
            "title": "Description",
            "type": "string"
          },
          "reason": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Reason",
            "type": "string"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "enum": [
              "listing",
              "user",
              "message",
              "seller_header"
            ],
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "target_type",
          "target_id",
          "reason"
        ],
        "title": "ReportRequest",
        "type": "object"
      },
      "ReviewRequest": {
        "properties": {
          "rating": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "type": "integer"
          },
          "text": {
            "default": "",
            "maxLength": 2000,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "rating"
        ],
        "title": "ReviewRequest",
        "type": "object"
      },
...
      "app__domains__fulfillment__schemas__ChallengeResponse": {
        "properties": {
          "challenge_id": {
            "format": "uuid",
            "title": "Challenge Id",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "purpose": {
            "enum": [
              "handover",
              "return"
            ],
            "title": "Purpose",
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "code",
          "purpose",
          "expires_at"
        ],
        "title": "ChallengeResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "4by4 For Hire API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/listings/{listing_id}/moderate": {
      "post": {
        "operationId": "moderate_listing_api_v1_admin_listings__listing_id__moderate_post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Moderate Listing Api V1 Admin Listings  Listing Id  Moderate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Moderate Listing",
        "tags": [
          "fulfillment",
          "trust"
        ]
      }
    },
    "/api/v1/admin/reports": {
      "get": {
        "operationId": "admin_reports_api_v1_admin_reports_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminReportResponse"
                  },
                  "title": "Response Admin Reports Api V1 Admin Reports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Reports",
        "tags": [
          "fulfillment",
          "trust"
        ]
      }
    },
...
    "/api/v1/bookings/{booking_id}/challenge/confirm": {
      "post": {
        "operationId": "confirm_challenge_api_v1_bookings__booking_id__challenge_confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Challenge",
        "tags": [
          "fulfillment",
          "trust"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/condition-reports": {
      "post": {
        "operationId": "condition_report_api_v1_bookings__booking_id__condition_reports_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConditionReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Condition Report Api V1 Bookings  Booking Id  Condition Reports Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Condition Report",
        "tags": [
          "fulfillment",
          "trust"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/disputes": {
      "post": {
        "operationId": "create_dispute_api_v1_bookings__booking_id__disputes_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Create Dispute Api V1 Bookings  Booking Id  Disputes Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Dispute",
        "tags": [
          "fulfillment",
          "trust"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/fulfillment": {
      "put": {
        "operationId": "schedule_fulfillment_api_v1_bookings__booking_id__fulfillment_put",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FulfillmentScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FulfillmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Schedule Fulfillment",
        "tags": [
          "fulfillment",
          "trust"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/inspection/accept": {
      "post": {
        "operationId": "accept_inspection_api_v1_bookings__booking_id__inspection_accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept Inspection",
        "tags": [
          "fulfillment",
          "trust"
        ]
      }
    },
...
    "/api/v1/bookings/{booking_id}/payment-at-pickup/acknowledge": {
      "post": {
        "operationId": "acknowledge_payment_api_v1_bookings__booking_id__payment_at_pickup_acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentAcknowledgementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "boolean"
                      }
                    ]
                  },
                  "title": "Response Acknowledge Payment Api V1 Bookings  Booking Id  Payment At Pickup Acknowledge Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Acknowledge Payment",
        "tags": [
          "fulfillment",
          "trust"
        ]
      }
    },
...
    "/api/v1/bookings/{booking_id}/return/initiate": {
      "post": {
        "operationId": "initiate_return_api_v1_bookings__booking_id__return_initiate_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Initiate Return",
        "tags": [
          "fulfillment",
          "trust"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/reviews": {
      "post": {
        "operationId": "create_review_api_v1_bookings__booking_id__reviews_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "integer"
                      }
                    ]
                  },
                  "title": "Response Create Review Api V1 Bookings  Booking Id  Reviews Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Review",
        "tags": [
          "fulfillment",
          "trust"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/{purpose}/challenge": {
      "post": {
        "operationId": "create_challenge_api_v1_bookings__booking_id___purpose__challenge_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "purpose",
            "required": true,
            "schema": {
              "title": "Purpose",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__domains__fulfillment__schemas__ChallengeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Challenge",
        "tags": [
          "fulfillment",
          "trust"
        ]
      }
    },
...
    "/api/v1/reports": {
      "post": {
        "operationId": "create_report_api_v1_reports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Create Report Api V1 Reports Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Report",
        "tags": [
          "fulfillment",
          "trust"
        ]
      }
    },
...
  }
}